+++ sample/src/main/java/com/codetroopers/betterpickers/sample/activity/calendardatepicker/SampleCalendarDateThemeCustom.java
@@ -31,6 +31,8 @@ public class SampleCalendarDateThemeCustom extends BaseSampleActivity
             @Override
             public void onClick(View v) {
                 CalendarDatePickerDialogFragment cdp = new CalendarDatePickerDialogFragment()
+                        .setDoneText(getString(R.string.button_label_custom_ok))
+                        .setCancelText(getString(R.string.button_label_custom_cancel))
                         .setThemeCustom(R.style.MyCustomBetterPickersRadialTimePickerDialog);
                 cdp.show(getSupportFragmentManager(), FRAG_TAG_DATE_PICKER);
             }